+++ apps/web/src/components/Composer/LinkPreviews.tsx
@@ -5,7 +5,7 @@ import { XMarkIcon } from '@heroicons/react/24/outline';
 import { KNOWN_ATTRIBUTES } from '@hey/data/constants';
 import getURLs from '@hey/helpers/getURLs';
 import { MetadataAttributeType } from '@lens-protocol/metadata';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { usePublicationAttachmentStore } from 'src/store/non-persisted/publication/usePublicationAttachmentStore';
 import { usePublicationAttributesStore } from 'src/store/non-persisted/publication/usePublicationAttributesStore';
 import { usePublicationStore } from 'src/store/non-persisted/publication/usePublicationStore';
@@ -23,7 +23,6 @@ const LinkPreviews: FC<LinkPreviewProps> = ({
   const { attachments } = usePublicationAttachmentStore((state) => state);
   const { addAttribute, getAttribute, removeAttribute } =
     usePublicationAttributesStore();
-  const [showRemove, setShowRemove] = useState(false);
 
   const urls = getURLs(publicationContent);
 
@@ -46,7 +45,6 @@ const LinkPreviews: FC<LinkPreviewProps> = ({
   return (
     <div className="relative m-5">
       <Oembed
-        onLoad={(og) => setShowRemove(og?.title ? true : false)}
         openActionEmbed={openActionEmbed}
         openActionEmbedLoading={openActionEmbedLoading}
         url={urls[0]}
