+++ apps/web/src/components/Publication/OpenAction/OnBody.tsx
@@ -36,10 +36,7 @@ const OpenActionOnBody: FC<OpenActionOnBodyProps> = ({ publication }) => {
       {isFeatureAvailable('rent-ads') &&
         module.contract.address ===
           VerifiedOpenActionModules.RentableBillboard && (
-          <RentableBillboardOpenAction
-            module={module as UnknownOpenActionModuleSettings}
-            publication={publication}
-          />
+          <RentableBillboardOpenAction />
         )}
     </div>
   );
