+++ apps/web/src/components/Shared/Navbar/MenuItems.tsx
@@ -3,7 +3,7 @@ import type { FC } from 'react';
 import Link from 'next/link';
 import { useProfileStore } from 'src/store/persisted/useProfileStore';
 
-import LoginButton from './LoginButton';
+import LoginButton from '../LoginButton';
 import SignedUser from './SignedUser';
 import SignupButton from './SignupButton';
