+++ apps/web/src/components/Staff/Users/Overview/Tool/OnchainIdentities.tsx
@@ -1,7 +1,7 @@
 import type { ProfileOnchainIdentity } from '@hey/lens';
 import type { FC } from 'react';
 
-import MetaDetails from '@components/Shared/Staff/MetaDetails';
+import MetaDetails from '@components/Shared/MetaDetails';
 import {
   GlobeAltIcon,
   HashtagIcon,
