+++ apps/web/src/components/StaffTools/Panels/Publication.tsx
@@ -1,7 +1,7 @@
 import type { MirrorablePublication } from '@hey/lens';
 import type { FC } from 'react';
 
-import MetaDetails from '@components/Shared/Staff/MetaDetails';
+import MetaDetails from '@components/Shared/MetaDetails';
 import {
   HashtagIcon,
   RectangleStackIcon,
